Made higlass_checks files, added files_not_registered_with_higlass to it. Also made and scheduled generate_higlass_view_confs_files and associated action for Chad to work on
